catalogdb.py: edit model for catalog_ver25_to_ver31_full_unique and catalog_ver25_to_ver31_full_all (#208)

+++ python/sdssdb/peewee/sdss5db/catalogdb.py
@@ -98,7 +98,7 @@ class Catalog(CatalogdbModel):
 
 class Catalog_ver25_to_ver31_full_unique(CatalogdbModel):
 
-    id = BigIntegerField(primary_key=True)
+    id = IntegerField(primary_key=True)
 
     class Meta:
         table_name = 'catalog_ver25_to_ver31_full_unique'
@@ -106,7 +106,7 @@ class Catalog_ver25_to_ver31_full_unique(CatalogdbModel):
 
 class Catalog_ver25_to_ver31_full_all(CatalogdbModel):
 
-    id = BigIntegerField(primary_key=True)
+    id = IntegerField(primary_key=True)
 
     class Meta:
         table_name = 'catalog_ver25_to_ver31_full_all'
